Add support for different upstream port in different lambdas.
In testing, I find a bug that when restoring another lambda in net
namespace, CRIU will complain that "Address already in use". The reason
is that send KILL signal is async in linux kernel and when restoring the
old python process's resource still alive (at least its socket is).

So I choose different port in different lambda, then when switching it
is guaranteed that it will use different port so the above problem might
be solved.